+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1 (2016-06-29)
+- Allow more Data into create card authorization
+- Update seal calcul method
+
 ## 1.0.0 (2016-06-28)
 - Allow more Data into create card authorization